Benchmark bigger payloads.
Most of the previous benchmarks worked on rather small data.  In
particular, the benchmarks that used larger Vectors did not compare
performance with binary and cereal.

This commit changes this, so that we can compare the libraries for large
payloads.

It it interesting to note that it seems decoding is only faster in store
than in cereal if we have larger payloads.