Compatibility with LTS-3.13.
Added stack-lts-3.13.yaml, which builds with LTS-Haskell 3.13, and added
dependencies.

Also, removed the lower version bounds for base-orphans, th-reify-many,
and th-orphans.  Those are needed in order to build with ghc-7.8, using
stack-7.8.yaml.  Since versions are fixed in stack-7.8.yaml anyway, we
don't need the bounds in the .cabal file.